우진 상품디테일

+++ src/main/java/com/green/winey_final/detail/model/SelCountVo.java
@@ -6,6 +6,6 @@ import lombok.Data;
 
 public class SelCountVo {
     private Long productId;
-    private Long reviewLevel;
+    private Integer reviewLevel;
 
 }
